webapp/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
azure-identity==1.15.0
azure-mgmt-resource==23.1.1
azure-mgmt-compute==30.4.0
azure-mgmt-storage==21.1.0
azure-mgmt-sql==3.0.1
azure-mgmt-web==7.2.0
requests==2.31.0
gunicorn==21.2.0